uftpd.py: New attempt to avoid false command client closing
Looks more stable now with Nautilus.
It seems that if during after data tranfer the data session ist first closed by
the client, the command session is signalled that it was closed (not the data session)
In that case the sever closes the command session, which is not the intention.
The approach to avoid this situation is closing the data session earlier,
but that is not guaranteed to work. In the worst case, the command session has to
be re-opened, what file managers will do.