Use separate server sockets per interface
For the PASV command the server needs to know which
interface (AP_IF or STA_IF) the client connected to.
Since usocket lacks the getsockname() method, we need
to use separate FTP server sockets per interface and
use a lambda function to pass the local address to
accept_ftp_connect().