ftp_pycom.py: Variant for Pycom devices
Pycom devices have a good working ftp server built-in. This is just
an option e.g. for situations when an active connection is needed, or
e.g. for use with the Chrome browser.
It starts by default on port 5024 (instead of 21), and terminates
if the client send a quit command.